「Spring Initializrでプロジェクトを作成、ダウンロード」「Eclipseプロジェクトを作成、動作確認」セクションを書きました。また、動作確認を行うためにソースコードを修正しました。
