+++ .gitignore
@@ -1,6 +1,7 @@
 .gradle
 /build/
 !gradle/wrapper/gradle-wrapper.jar
+*.log
 
 ### STS ###
 .apt_generated
