+++ src/main/java/me/u6k/sample_spring_boot_with_gradle/Main.java
@@ -1,0 +1,14 @@
+
+package me.u6k.sample_spring_boot_with_gradle;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Main {
+
+    public static void main(String[] args) {
+        SpringApplication.run(Main.class, args);
+    }
+
+}
