--- src/main/java/me/u6k/samplespringbootwithgradle/SampleSpringBootWithGradleApplication.java
@@ -1,12 +1,0 @@
-package me.u6k.samplespringbootwithgradle;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SampleSpringBootWithGradleApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SampleSpringBootWithGradleApplication.class, args);
-	}
-}
