+++ src/main/resources/application.properties
@@ -1,0 +1,10 @@
+spring.datasource.driverClassName=org.apache.derby.jdbc.EmbeddedDriver
+spring.datasource.url=jdbc:derby:${APP_DB_PATH:build/db/};create=true
+spring.datasource.username=root
+spring.datasource.password=root
+spring.jpa.hibernate.show-sql=true
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.database-platform=org.hibernate.dialect.DerbyDialect
+
+logging.level.root=INFO
+logging.level.me.u6k=DEBUG
