+++ src/test/java/me/u6k/sample_spring_boot_with_gradle/MainTest.java
@@ -1,0 +1,17 @@
+
+package me.u6k.sample_spring_boot_with_gradle;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class MainTest {
+
+    @Test
+    public void contextLoads() {
+    }
+
+}
