--- src/test/java/me/u6k/samplespringbootwithgradle/SampleSpringBootWithGradleApplicationTests.java
@@ -1,16 +1,0 @@
-package me.u6k.samplespringbootwithgradle;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class SampleSpringBootWithGradleApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
